# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Usuarios",
        "description": "Usuarios",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listado de todos los registros de los usuarios",
                "operationId": "4db02f0d9ccd97469981e0cb03d1a41e",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Bearer token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listar los usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Aderson Felix"
                                                    },
                                                    "lastName": {
                                                        "type": "string",
                                                        "example": "Jara Lazaro"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "Jara@gmail.com"
                                                    },
                                                    "password": {
                                                        "type": "string",
                                                        "example": "123456"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "Administrador"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Iniciar sesión",
                "description": "Inicia sesión de un usuario",
                "operationId": "2cbb16711380c1958ba74271ad13dcd5",
                "requestBody": {
                    "description": "Credenciales de inicio de sesión",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "Jara@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "asdasdasdasdasdasdasd"
                                        },
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "email": {
                                                        "type": "string",
                                                        "example": "Jara@gmail.com"
                                                    },
                                                    "password": {
                                                        "type": "string",
                                                        "example": "123456"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Credenciales inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "invalid_credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear el token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "could_not_create_token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/create": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear usuario",
                "operationId": "7f4ec36156a2c5098689d2c2a0771602",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Bearer token"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Body para crear usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Luis"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Rojas"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "Jara@gmail"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "Administrador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Listar los usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Aderson Felix"
                                                    },
                                                    "lastName": {
                                                        "type": "string",
                                                        "example": "Jara Lazaro"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "Jara@gmail.com"
                                                    },
                                                    "password": {
                                                        "type": "string",
                                                        "example": "123456"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "Administrador"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar usuario",
                "operationId": "a658c3493d56ef977986fc38a6518c32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Bearer token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listar los usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/": {
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Modificar Usuario",
                "operationId": "71bc8a2106a96facb0d28dee6f79a6f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Bearer token"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Body para crear usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Luis"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Rojas"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "Jara@gmail"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "Administrador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Listar los usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}